Se crean controladores abstractos y errores

+++ src/module/cars/repository/carRepository.js
@@ -1,6 +1,9 @@
 const { fromDbToEntity } = require('../mapper/carMapper')
+const CarIdNotDefinedError = require('./error/carIdNotDefinedError')
+const CarNotFoundError = require('./error/carNotFoundError')
+const AbstractCarRepository = require('./abstractCarRepository')
 
-module.exports = class Repository {
+module.exports = class Repository extends AbstractCarRepository {
 
     /**
      * 
@@ -8,23 +11,52 @@ module.exports = class Repository {
      */
 
     constructor(carModel) {
+        super()
         this.carModel = carModel
     }
 
     async getData() {
         const carList = await this.carModel.findAll()
-        return fromDbToEntity(carList)
+        return carList.map(fromDbToEntity)
     }
+
+    /**
+     * 
+     * @param {import('../entity/Car')} car 
+     */
+
     async save(car) {
-        console.log(car)
         const options = { isNewRecord: !car.id }
         const newCar = await this.carModel.build(car, options)
         newCar.save()
     }
+
+    /**
+     * 
+     * @param {Number} id 
+     */
+
     async getById(id) {
-        return await this.carModel.findByPk(id)
+        const car = await this.carModel.findByPk(id)
+
+        if (!car) {
+            throw new CarNotFoundError(`No se encontro un auto con id ${id}`)
+        }
+
+        return fromDbToEntity(car)
     }
+
+    /**
+     * 
+     * @param {Number} id 
+     */
+
     async remove(id) {
+
+        if (!id) {
+            throw new CarIdNotDefinedError(`No se encontro id`)
+        }
+
         return await this.carModel.destroy({
             where: {
                 id: id
