+++ src/module/cars/service/carService.js
@@ -1,3 +1,6 @@
+const CarNotDefinedError = require('./error/carNotDefined')
+const CarIdNotDefined = require('./error/carIdNotDefined')
+
 module.exports = class Service {
     constructor(carRepository) {
         this.carRepository = carRepository
@@ -8,12 +11,21 @@ module.exports = class Service {
     }
 
     async save(car) {
+        if (car === undefined) {
+            throw new CarNotDefinedError()
+        }
         return this.carRepository.save(car)
     }
     async getById(id) {
+        if (id === undefined) {
+            throw new CarIdNotDefinedError()
+        }
         return this.carRepository.getById(id)
     }
-    async remove(id){
+    async remove(id) {
+        if (id === undefined) {
+            throw new CarIdNotDefinedError()
+        }
         return this.carRepository.remove(id)
     }
 }
